+++ task_planner/src/main/java/com/jurgen/task_planner/services/IIssueService.java
@@ -9,8 +9,8 @@ import com.jurgen.task_planner.models.requests.UpdateIssueRequest;
 
 public interface IIssueService {
     List<IssueDto> getIssuesForTenant(int tenantId);
-    IssueDto getIssue(int tenantId, long issueId) throws HttpErrorException;
+    IssueDto getIssue(int tenantId, int issueId) throws HttpErrorException;
     IssueDto createIssue(int tenantId, CreateIssueRequest request) throws HttpErrorException;
-    IssueDto updateIssue(int tenantId, long issueId, UpdateIssueRequest request) throws HttpErrorException;
-    void deleteIssue(int tenantId, long issueId) throws HttpErrorException;
+    IssueDto updateIssue(int tenantId, int issueId, UpdateIssueRequest request) throws HttpErrorException;
+    void deleteIssue(int tenantId, int issueId) throws HttpErrorException;
 }
